+++ src/Component/OpenIdConnect/OpenIdConnectToken.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Heptacom\AdminOpenAuth\Component\OpenIdConnect;
 
+use Jose\Component\Signature\Serializer\CompactSerializer;
+use Jose\Component\Signature\Serializer\JWSSerializerManager;
 use Shopware\Core\Framework\Struct\Struct;
 
 final class OpenIdConnectToken extends Struct
@@ -12,11 +14,22 @@ final class OpenIdConnectToken extends Struct
 
     protected ?int $expires_in = null;
 
+    protected ?string $id_token = null;
+
     protected ?int $refresh_expires_in = null;
 
     protected ?string $refresh_token = null;
 
     protected ?string $token_type = '';
+
+    private JWSSerializerManager $serializerManager;
+
+    public function __construct()
+    {
+        $this->serializerManager = new JWSSerializerManager([
+            new CompactSerializer(),
+        ]);
+    }
 
     public function getAccessToken(): ?string
     {
@@ -36,6 +49,36 @@ final class OpenIdConnectToken extends Struct
     public function setExpiresIn(?int $expires_in): void
     {
         $this->expires_in = $expires_in;
+    }
+
+    public function getIdToken(): ?string
+    {
+        return $this->id_token;
+    }
+
+    public function setIdToken(?string $id_token): OpenIdConnectToken
+    {
+        $this->id_token = $id_token;
+        return $this;
+    }
+
+    public function getIdTokenPayload(): ?array
+    {
+        if ($this->id_token === null) {
+            return null;
+        }
+
+        $idTokenPayload = $this->serializerManager->unserialize($this->id_token)->getPayload();
+        if ($idTokenPayload === null) {
+            return [];
+        }
+
+        $decodedPayload = \json_decode($idTokenPayload, true, 512, JSON_THROW_ON_ERROR);
+        if (!\is_array($decodedPayload)) {
+            throw new \InvalidArgumentException('Invalid id_token payload. Expected JSON object.', 1739779042);
+        }
+
+        return $decodedPayload;
     }
 
     public function getRefreshExpiresIn(): ?int